room

+++ src/main/java/com/da/killerwebsite/repository/ScriptRepository.java
@@ -3,5 +3,5 @@ package com.da.killerwebsite.repository;
 import com.da.killerwebsite.entity.Script;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ScriptRepository extends JpaRepository<Script, Long> {
+public interface ScriptRepository extends JpaRepository<Script, Integer> {
 }
